--- a/business/dashboards.go
+++ b/business/dashboards.go
@@ -2,6 +2,7 @@ package business
 
 import (
 	"fmt"
+	"sort"
 	"strings"
 	"sync"
 
@@ -11,32 +12,90 @@ import (
 	"github.com/kiali/kiali/log"
 	"github.com/kiali/kiali/models"
 	"github.com/kiali/kiali/prometheus"
+	"github.com/kiali/kiali/prometheus/internalmetrics"
 )
 
 // DashboardsService deals with fetching dashboards from k8s client
 type DashboardsService struct {
-	prom prometheus.ClientInterface
-	mon  kubernetes.KialiMonitoringInterface
+	prom      prometheus.ClientInterface
+	k8sClient kubernetes.KialiMonitoringInterface
 }
 
 // NewDashboardsService initializes this business service
-func NewDashboardsService(mon kubernetes.KialiMonitoringInterface, prom prometheus.ClientInterface) DashboardsService {
-	return DashboardsService{prom: prom, mon: mon}
+func NewDashboardsService(k8sClient kubernetes.KialiMonitoringInterface, prom prometheus.ClientInterface) DashboardsService {
+	return DashboardsService{prom: prom, k8sClient: k8sClient}
 }
 
-func (in *DashboardsService) loadDashboardResource(namespace, template string) (*v1alpha1.MonitoringDashboard, error) {
+func (in *DashboardsService) k8s() (kubernetes.KialiMonitoringInterface, error) {
+	// Lazy init
+	if in.k8sClient == nil {
+		client, err := kubernetes.NewKialiMonitoringClient()
+		if err != nil {
+			return nil, fmt.Errorf("Cannot initialize Kiali Monitoring Client: %v", err)
+		}
+		in.k8sClient = client
+	}
+	return in.k8sClient, nil
+}
+
+// k8sGetDashboard wraps KialiMonitoringInterface.GetDashboard with client creation & error handling
+func (in *DashboardsService) k8sGetDashboard(namespace, template string) (*v1alpha1.MonitoringDashboard, error) {
+	client, err := in.k8s()
+	if err != nil {
+		return nil, err
+	}
+	return client.GetDashboard(namespace, template)
+}
+
+// k8sGetDashboards wraps KialiMonitoringInterface.GetDashboards with client creation & error handling
+func (in *DashboardsService) k8sGetDashboards(namespace string) ([]v1alpha1.MonitoringDashboard, error) {
+	client, err := in.k8s()
+	if err != nil {
+		return nil, err
+	}
+	return client.GetDashboards(namespace)
+}
+
+func (in *DashboardsService) loadRawDashboardResource(namespace, template string) (*v1alpha1.MonitoringDashboard, error) {
 	// There is an override mechanism with dashboards: default dashboards can be provided in Kiali namespace,
 	// and can be overriden in app namespace.
 	// So we look for the one in app namespace first, and only if not found fallback to the one in istio-system.
-	dashboard, err := in.mon.GetDashboard(namespace, template)
+	dashboard, err := in.k8sGetDashboard(namespace, template)
 	if err != nil {
 		cfg := config.Get()
-		dashboard, err = in.mon.GetDashboard(cfg.IstioNamespace, template)
+		dashboard, err = in.k8sGetDashboard(cfg.IstioNamespace, template)
 		if err != nil {
 			return nil, err
 		}
 	}
 	return dashboard, err
+}
+
+func (in *DashboardsService) loadRawDashboardResources(namespace string) (map[string]v1alpha1.MonitoringDashboard, error) {
+	all := make(map[string]v1alpha1.MonitoringDashboard)
+
+	// From Kiali namespace
+	cfg := config.Get()
+	dashboards, err := in.k8sGetDashboards(cfg.IstioNamespace)
+	if err != nil {
+		return nil, err
+	}
+	for _, d := range dashboards {
+		all[d.Name] = d
+	}
+
+	// From specific namespace
+	if namespace != cfg.IstioNamespace {
+		dashboards, err = in.k8sGetDashboards(namespace)
+		if err != nil {
+			return nil, err
+		}
+		for _, d := range dashboards {
+			all[d.Name] = d
+		}
+	}
+
+	return all, nil
 }
 
 func (in *DashboardsService) loadAndResolveDashboardResource(namespace, template string, loaded map[string]bool) (*v1alpha1.MonitoringDashboard, error) {
@@ -45,7 +104,7 @@ func (in *DashboardsService) loadAndResolveDashboardResource(namespace, template
 		return nil, fmt.Errorf("Cannot load dashboard %s due to circular dependency detected. Already loaded dependencies: %v", template, loaded)
 	}
 	loaded[template] = true
-	dashboard, err := in.loadDashboardResource(namespace, template)
+	dashboard, err := in.loadRawDashboardResource(namespace, template)
 	if err != nil {
 		return nil, err
 	}
@@ -220,6 +279,45 @@ func (in *DashboardsService) GetIstioDashboard(params prometheus.IstioMetricsQue
 	return &dashboard, nil
 }
 
+// GetCustomDashboardRefs finds all dashboard IDs and Titles associated to this app and add them to the model
+func (in *DashboardsService) GetCustomDashboardRefs(namespace, app, version string, pods models.Pods) []models.Runtime {
+	var err error
+	promtimer := internalmetrics.GetGoFunctionMetric("business", "DashboardsService", "GetCustomDashboardRefs")
+	defer promtimer.ObserveNow(&err)
+
+	uniqueRefsList := getUniqueRuntimes(pods)
+	if len(uniqueRefsList) > 0 {
+		return in.buildRuntimesList(namespace, uniqueRefsList)
+	}
+	if app != "" {
+		// Discovery only works with proper "app" labelling
+		return in.discoverRuntimesList(namespace, app, version)
+	}
+	return []models.Runtime{}
+}
+
+func getUniqueRuntimes(pods models.Pods) []string {
+	// Get uniqueness from plain list rather than map to preserve ordering; anyway, very low amount of objects is expected
+	uniqueRefs := []string{}
+	for _, pod := range pods {
+		for _, ref := range pod.RuntimesAnnotation {
+			if ref != "" {
+				exists := false
+				for _, existingRef := range uniqueRefs {
+					if ref == existingRef {
+						exists = true
+						break
+					}
+				}
+				if !exists {
+					uniqueRefs = append(uniqueRefs, ref)
+				}
+			}
+		}
+	}
+	return uniqueRefs
+}
+
 func (in *DashboardsService) buildRuntimesList(namespace string, templatesNames []string) []models.Runtime {
 	dashboards := make([]*v1alpha1.MonitoringDashboard, len(templatesNames))
 	wg := sync.WaitGroup{}
@@ -227,7 +325,7 @@ func (in *DashboardsService) buildRuntimesList(namespace string, templatesNames 
 	for idx, template := range templatesNames {
 		go func(i int, tpl string) {
 			defer wg.Done()
-			dashboard, err := in.loadDashboardResource(namespace, tpl)
+			dashboard, err := in.loadRawDashboardResource(namespace, tpl)
 			if err != nil {
 				log.Errorf("Cannot get dashboard %s in namespace %s. Error was: %v", tpl, namespace, err)
 			} else {
@@ -243,26 +341,98 @@ func (in *DashboardsService) buildRuntimesList(namespace string, templatesNames 
 		if dashboard == nil {
 			continue
 		}
-		runtime := dashboard.Spec.Runtime
-		ref := models.DashboardRef{
-			Template: dashboard.Name,
-			Title:    dashboard.Spec.Title,
-		}
-		found := false
-		for i := range runtimes {
-			rtObj := &runtimes[i]
-			if rtObj.Name == runtime {
-				rtObj.DashboardRefs = append(rtObj.DashboardRefs, ref)
-				found = true
-				break
+		runtimes = addDashboardToRuntimes(dashboard, runtimes)
+	}
+	return runtimes
+}
+
+func (in *DashboardsService) fetchMetricNames(namespace, app, version string) []string {
+	labels := fmt.Sprintf(`{namespace="%s",app="%s"`, namespace, app)
+	if version != "" {
+		labels += fmt.Sprintf(`,version="%s"`, version)
+	}
+	labels += "}"
+	metrics, err := in.prom.GetMetricsForLabels([]string{labels})
+	if err != nil {
+		log.Errorf("Runtimes discovery failed, cannot load metrics for labels: %s. Error was: %v", labels, err)
+	}
+	return metrics
+}
+
+func (in *DashboardsService) discoverRuntimesList(namespace, app, version string) []models.Runtime {
+	var metrics []string
+	wg := sync.WaitGroup{}
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		metrics = in.fetchMetricNames(namespace, app, version)
+	}()
+
+	allDashboards, err := in.loadRawDashboardResources(namespace)
+	if err != nil {
+		log.Errorf("Runtimes discovery failed, cannot load dashboards in namespace %s. Error was: %v", namespace, err)
+		return []models.Runtime{}
+	}
+
+	wg.Wait()
+	return runDiscoveryMatcher(metrics, allDashboards)
+}
+
+func runDiscoveryMatcher(metrics []string, allDashboards map[string]v1alpha1.MonitoringDashboard) []models.Runtime {
+	// In all dashboards, finds the ones that match the metrics set
+	// We must exclude from the results included dashboards when both the including and the included dashboards are matching
+	runtimesMap := make(map[string]*v1alpha1.MonitoringDashboard)
+	for _, d := range allDashboards {
+		dashboard := d // sticky reference
+		matchReference := strings.TrimSpace(dashboard.Spec.DiscoverOn)
+		if matchReference != "" {
+			for _, metric := range metrics {
+				if matchReference == strings.TrimSpace(metric) {
+					if _, exists := runtimesMap[dashboard.Name]; !exists {
+						runtimesMap[dashboard.Name] = &dashboard
+					}
+					// Mark included dashboards as already found
+					// and set them "nil" to not show them as standalone dashboards even if they match
+					for _, item := range dashboard.Spec.Items {
+						if item.Include != "" {
+							runtimesMap[item.Include] = nil
+						}
+					}
+					break
+				}
 			}
 		}
-		if !found {
-			runtimes = append(runtimes, models.Runtime{
-				Name:          runtime,
-				DashboardRefs: []models.DashboardRef{ref},
-			})
+	}
+	runtimes := []models.Runtime{}
+	for _, dashboard := range runtimesMap {
+		if dashboard != nil {
+			runtimes = addDashboardToRuntimes(dashboard, runtimes)
 		}
+	}
+	sort.Slice(runtimes, func(i, j int) bool { return runtimes[i].Name < runtimes[j].Name })
+	return runtimes
+}
+
+func addDashboardToRuntimes(dashboard *v1alpha1.MonitoringDashboard, runtimes []models.Runtime) []models.Runtime {
+	runtime := dashboard.Spec.Runtime
+	ref := models.DashboardRef{
+		Template: dashboard.Name,
+		Title:    dashboard.Spec.Title,
+	}
+	found := false
+	for i := range runtimes {
+		rtObj := &runtimes[i]
+		if rtObj.Name == runtime {
+			rtObj.DashboardRefs = append(rtObj.DashboardRefs, ref)
+			found = true
+			break
+		}
+	}
+	if !found {
+		runtimes = append(runtimes, models.Runtime{
+			Name:          runtime,
+			DashboardRefs: []models.DashboardRef{ref},
+		})
 	}
 	return runtimes
 }