--- a/deploy/dashboards/go.yaml
+++ b/deploy/dashboards/go.yaml
@@ -5,6 +5,7 @@ metadata:
 spec:
   title: Go Metrics
   runtime: Go
+  discoverOn: "go_info"
   items:
   - chart:
       name: "Threads"