--- a/deploy/dashboards/micrometer-1.0.6-jvm-pool.yaml
+++ b/deploy/dashboards/micrometer-1.0.6-jvm-pool.yaml
@@ -5,6 +5,7 @@ metadata:
 spec:
   runtime: JVM
   title: JVM Pool Metrics
+  discoverOn: "jvm_buffer_total_capacity_bytes"
   items:
   - chart:
       name: "Pool buffer memory used"