--- a/deploy/dashboards/micrometer-1.0.6-jvm.yaml
+++ b/deploy/dashboards/micrometer-1.0.6-jvm.yaml
@@ -5,6 +5,7 @@ metadata:
 spec:
   runtime: JVM
   title: JVM Metrics
+  discoverOn: "jvm_threads_live"
   items:
   - chart:
       name: "Total live threads"