… business/dashboards_test.go
… deploy/dashboards/micrometer-1.1-jvm.yaml
… deploy/dashboards/microprofile-1.1.yaml
… deploy/dashboards/microprofile-x.y.yaml
… deploy/dashboards/nodejs.yaml
… deploy/dashboards/thorntail.yaml
… deploy/dashboards/tomcat.yaml
… deploy/dashboards/vertx-client.yaml
… deploy/dashboards/vertx-eventbus.yaml
… deploy/dashboards/vertx-pool.yaml
… deploy/dashboards/vertx-server.yaml
… deploy/kubernetes/clusterrole.yaml
… deploy/openshift/clusterrole.yaml
… operator/roles/kiali/templates/dashboards/go.yaml
… operator/roles/kiali/templates/dashboards/micrometer-1.0.6-jvm-pool.yaml
… operator/roles/kiali/templates/dashboards/micrometer-1.0.6-jvm.yaml
… operator/roles/kiali/templates/dashboards/micrometer-1.1-jvm.yaml
… operator/roles/kiali/templates/dashboards/microprofile-1.1.yaml
… operator/roles/kiali/templates/dashboards/microprofile-x.y.yaml
… operator/roles/kiali/templates/dashboards/nodejs.yaml
… operator/roles/kiali/templates/dashboards/thorntail.yaml
… operator/roles/kiali/templates/dashboards/tomcat.yaml
… operator/roles/kiali/templates/dashboards/vertx-client.yaml
… operator/roles/kiali/templates/dashboards/vertx-eventbus.yaml
… operator/roles/kiali/templates/dashboards/vertx-pool.yaml
… operator/roles/kiali/templates/dashboards/vertx-server.yaml