WIP / draft.
PR is based on KIALI-2488 / https://github.com/kiali/kiali/pull/959 which should be merged first.

Some performance consideration, measuring the loading time of Workload Details page:
- Before this PR, with no runtimes dashboard: avg ~= 13ms
- Before this PR, with runtimes dashboards: avg ~= 20ms
- After this PR, with  runtimes dashboards: avg ~= 20ms (no change, as expected)
- After this PR, with no dashboard: avg ~= 180ms

This peak due to discovery was kinda expected, see description in https://issues.jboss.org/browse/KIALI-2546 ; suggestion is to cache dashboards.

Note however that the increased latency *only* affects the Workloads / App details page, (not list pages), which makes it less critical as there's no scaling issue (we only show workload details one by one).