@rhqci 2019-04-01
Jenkins CI: kiali-core-pr-e2e-test [#765](https://paste.fedoraproject.org/paste/eX9we6qi6Fyl5c42nOIAqQ)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1570]](https://paste.fedoraproject.org/paste/2UuiIZ~yyXt-P~wVF4H~-A)