@jotak 2019-04-01
Ideas to minimize impact on performances, non-exclusive:

- Caching dashboards (**pros**: efficient solution; **cons**: need to handle cache updates / invalidation)
- Flag-enabled feature (**pros**: can turn out completely to have 0 perf impact; **cons**: still perf issue when enabled)
- Async loading of available dashboards list (**pros**: fixes latency issue on Workload Details main info; **cons**: total loading time still unchanged (or even increased)