@rhqci 2019-04-01
Jenkins CI: kiali-core-pr-e2e-test [#766](https://paste.fedoraproject.org/paste/xWk5W2sN7uhIUYCLCMnUIA)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1571]](https://paste.fedoraproject.org/paste/yCkSqX58kWFY2hFrttIe~g)