@lucasponce 2019-04-02
> Note however that the increased latency only affects the Workloads / App details page, (not list pages), which makes it less critical as there's no scaling issue (we only show workload details one by one).

If this is under reasonable parameters, it doesn't sound bad.