@rhqci 2019-04-02
Jenkins CI: kiali-core-pr-e2e-test [#773](https://paste.fedoraproject.org/paste/h~c9oQNqyp~d0B4RMOMW~w)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1580]](https://paste.fedoraproject.org/paste/O0zN97S2b0GNE8m4Kugadw)