@rhqci 2019-04-03
Jenkins CI: kiali-core-pr-e2e-test [#786](https://paste.fedoraproject.org/paste/Gx3JNEDlYJTRXA0SFh8iOg)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1597]](https://paste.fedoraproject.org/paste/7RmTD5RYzMt~msEyS~1NEA)