@rhqci 2019-04-03
Jenkins CI: kiali-core-pr-e2e-test [#787](https://paste.fedoraproject.org/paste/ibMJixtVB1SXe8XbMJ0SRQ)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1598]](https://paste.fedoraproject.org/paste/7mXWYI~perYrkjPZrx0Gkw)