@rhqci 2019-04-03
Jenkins CI: kiali-core-pr-e2e-test [#788](https://paste.fedoraproject.org/paste/qGtaFz54OmCDzP6FwF-VqQ)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1599]](https://paste.fedoraproject.org/paste/ERLVjML51VtenRkFe3RlSw)