@skondkar 2019-04-04
Verified on:
kiali-ui 0.17.0 (0af3b25fe70da258149773a96468f8164db575aa)
kiali v0.17.0-SNAPSHOT (f0a2fb7775c46de9cf1a18272a90cf3cede5fd87)

Used bookinfo-runtimes and removed kiali.io/runtimes annotations from deplyment.yamls. After deploying,  all dashboards are discovered in Kiali with metrics.