@rhqci 2019-04-04
Jenkins CI: kiali-core-pr-e2e-test [#824](https://paste.fedoraproject.org/paste/SCTkaP9IA6n2k5rwgJc1zQ)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1645]](https://paste.fedoraproject.org/paste/iRcxgvX7oPZ3bJE97lW0PA)