@jotak 2019-04-04
So ... this is quite "funny". I'm seeing a race on the Config object, just because I've added a couple of `config.Get()`. The race makes sense, but what doesn't make sense to me is why it wasn't detected before? Because this PR doesn't change fundamentally anything here.

I'm currently seeing how to fix that... adding a mutex on Config works, but is costy. We should probably do something different and return a copy instead of pointer but i need to check all the code for impacts.
(cc @jmazzitelli @lucasponce )