@jmazzitelli 2019-04-04
> A RW semaphore would help ?
> 
> I guess W op only happens at init, right ?

Yes, that would be what I think, too. Add a RW lock with the W lock acquired at init time and released once the config is loaded. This is a * pointer being passed around for the config, but I wonder if there are things inside the config that are passed around by value that might be causing issues?? I'm just thinking out loud, but that's what I would be doing - see if a RW lock helps, but if this issue isn't happening at init time, then something else is going out, probably with the pass-by-ref vs. pass-by-value semantics in the config object itself.