@jotak 2019-04-04
@jmazzitelli the issue happens on `test-race` which, if I understand correctly, runs several tests concurrently. So in that context the "init time" occurs several times, not just once, I think...