@rhqci 2019-04-05
Jenkins CI: kiali-core-pr-e2e-test [#840](https://paste.fedoraproject.org/paste/RNcq83KtS1vSmDWZrohqJA)
 :red_circle: PRT Failed, Please Contact QE