@rhqci 2019-04-05
Jenkins CI: kiali-core-pr-e2e-test [#841](https://paste.fedoraproject.org/paste/An9WyVWdPvHc6-NGWiT3VA)
  * :heavy_check_mark: run-kiali-e2e-tests [#[1663]](https://paste.fedoraproject.org/paste/VT576UbfEFGBMLeOYnb0KQ)